+++ tests/error_origin.js
@@ -36,7 +36,7 @@ module.exports = async ({ test, equal, check }) => [
         // fails on ubuntu 18.04.4 node 12.18.3
         // check(inner_awaited.trace, t1 => !t1.includes(missing_async_msg))
         // check(inner_awaited.trace, t2 => t2.includes('error_origin.js'))
-        // console.log(`inner_awaited: ${JSON.stringify(inner_awaited)}\nia-trace: ${inner_awaited.trace}`)
+        // console.log(`inner_awaited: ${JSON.stringify(inner_awaited,null,2)}\nia-trace: ${inner_awaited.trace}`)
 
         check(non_awaited.trace, t3 => t3.includes(missing_async_msg))
         check(non_awaited.trace, t4 => !t4.includes('error_origin.js'))